chore: move singleton to creational pattern
