+++ creational/singleton/singleton.go
